+++ src/bootstrap.js
@@ -90,12 +90,13 @@ THREE.Bootstrap.prototype = {
     var aliases = _.extend({}, o.aliasdb, o.aliases);
 
     // Remove inline alias defs from plugins
-    plugins = _.filter(plugins, function (name) {
+    var filter = function (name) {
       var key = name.split(':');
       if (!key[1]) return true;
-      aliases[key[0]] = key[1];
+      aliases[key[0]] = [key[1]];
       return false;
-    });
+    };
+    plugins = _.filter(plugins, filter);
 
     // Unify arrays
     _.each(aliases, function (alias, key) {
@@ -105,6 +106,7 @@ THREE.Bootstrap.prototype = {
     // Look up aliases recursively
     function recurse(list, out, level) {
       if (level >= 256) throw "Plug-in alias recursion detected.";
+      list = _.filter(list, filter);
       _.each(list, function (name) {
         var alias = aliases[name];
         if (!alias) {